Initial implementation of per-segment prefetch request handling (#71113)
Introduces a new type of request that will be used for prefetching
individual segments. Requests are issued using the
Next-Router-Segment-Prefetch header. It accepts a path to the segment
that is being requested.

This feature part of PPR and will only be enabled when that flag is on.

Because these are prefetches, segment requests are always served from
the cache. They must never reach the application layer or invoke a
lambda. If a matching segment is not present in the cache, the server
responds with a 404 and an empty body. Segments that are fully dynamic
will also 404.

**In this PR, I have not yet implemented the actual segment
generation**; I'm using placeholder data for now. The smoke test I added
demonstrates that the placeholder data is correctly plumbed through the
build, ISR, and server layers to respond to a basic request.