Reject hanging promises (#70994)
When we prerender in `dynamicIO` the async APIs that provide access to
Request data return promises that never resolve. This is how we stall
the prerender in that particular slot while letting the rest of the App
continue. This is genernally fine however these promises can leak to
other contexts via setTimeout and after. After is particularly
concerning because it won't stop until the callbacks have finished and
if the callback is waiting on a promise that never resolves we end up
with a deadlock. To resolves this we now instrument these hanging
promises with an AbortSignal. When this signal aborts the promise
rejects. This works as long as the abort signal is always aborted at or
after when the prerender is aborted. In this implementation's case we
actually use the same abortSignal for both which provides the necessary
guarantee. I believe this is fine even when we synchronously abort a
render such as when you do something sync dynamic like call
`Math.random()`. If it turns out that we don't want to reject
immediately in these cases we can always switch to a second
AbortController that is only ever aborted after the prerender is
completed.