docs(javascript): fix wrong javascript code and a typo in `link.mdx` (#71143)
## 📚 Improving Documentation

Hello, I've fixed some incorrect javascript code and a typo in
[`link.mdx`](https://nextjs.org/docs/app/api-reference/components/link).

I made a few changes:

~~### 1. Changed `app/blog/page` to `app/blog/[slug]/page`~~

~~I believe it is accurate to include `[slug]` in the path.~~

### 2.  Added `export default`

The code block was missing `export default`, so I added it.

### 3. Corrected missing punctuation

### 4. Added `'use client`'

I think it is appropriate to add `'use client'` because the component
uses a client-side hook (`useIsAuthed`) that relies on client state,
which requires the component to be rendered on the client side.

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>