[dynamicIO] throw when prospective execute fails
When `dynamicIO` is enabled GET route handlers are called during build to see if they can be prerendered using similar semantics to page renders. This implementation uses a prospective render to warmup caches before it performs the final render to see if the output is entirely prerenderable.

One problem the current setup has is it swallows errors that happen during the prospective render which can make debugging certain failures hard, in particular if there are bugs within Next.js itself.

This change updates the implementation to throw if an error that is not related to interrupting a dynamic render is encountered during the prospective render.