Cache in CacheScopeStore in use cache (#71221)
Implements deduping of "use cache" entries using CacheScopeStore. This
is required for dynamic I/O.

We also have the option to use this for dynamic runtime deduping too,
rather than going to the underlying store and ensuring consistency
within the request. There are some negative tradeoffs with this approach
though that it might not be worth it for dynamic requests. For now, it's
up to the Cache Handler to do deduping if it wants to. I implemented
this in the default cache handler.

I tested this with a blank cache implementation since the existing cache
is microtasky. Since a working cache is required for other tests and
should be the default, I reverted the blank cache. So this doesn't have
the full test coverage yet in practice. Once we land configurable caches
we can use a blank cache in the newly added tests to get better
coverage.

I had to make the cache scope external since it's now shared with RSC
layer.