[dynamicIO] warn in debug mode when prospective renders error (#71266)
When prerendering pages and routes errors with dynamicIO enabled, errors
during the prospective render may be supressed from logging. Normally
this is expected and ok because the dynamic portion of a render may be
expected to error during prerendering. However some codepaths are not
hit the same between the prospective and final renders and so it can be
useful to have the ability to get additional insight into the runtime
behavior of the prospective render.

This change adds extra logging when `process.env.NEXT_DEBUG_BUILD` is
enabled. When this environment variable is set we log errors that throw
during the prospective render.