Set Next-Router-Stale-Time header from collected stale times (#71279)
That way the client router can use it to inform its cache.

This is only done during prerender which is typically the only things we
prefetch (other than loading.tsx).

I add it to `.meta` but this applies to both the RSC and HTML payloads I
believe. I'm not sure if there's away to only add it to the RSC payload.

For static parts of the HTML document we need a different strategy than
the head so this doesn't work for initial load.