Implement cacheScope by caching a Promise<CacheEntry> (#71286)
The cacheScope needs to contain the full CacheEntry, not just the
stream, so that it can transfer tag/life to the parent scope when it's
used.

I needed to do some refactoring. The tricky bit is that we want to store
a Promise into both the cacheHandler and the cacheScope as early as
possible to ensure they can dedupe and future requests while it's still
pending.