Encode FormData for cache keys using custom algorithm (#71341)
We need to encode FormData when we generate a complex cache key since
that's what RSC uses for encodeReply.

However, using the standard `multipart/form-data` encoding in `new
Response().arrayBuffer()` uses a large delimiter that's inefficient but
that delimiter is also randomly generated which completely defeats the
purpose of using it as a cache key.

Instead, we can use the size of each part as the delimiter.

It is important to something that is not just a character because
otherwise it's a potential security hole since someone could carefully
tailor input to cause cache poisoning.

E.g. if we used `:` as the delimiter then `[['a', 'b'],['c', 'd']]`
would be encoded as `a:b:c:d` but so would `[['a', 'b:c:d']]`.