[dynamicIO] track suspense boundaries only after excluding metadata and other boundaries
When we determine if a dynamicIO error representing a dynamic hold is allowed we use the parent component stack of the erroring component. Certain components have special marker compoennts int heir stack that allow us to make intelligent erroring and messaging decisions, in particular Metadata and related APIs. There was a bug in the current implementation where if there was a Suspense boundary parent of the metaata outlet it would be categorized as having something dynamic inside a Suspense boundary. This is erroneous becuase the MetaataOutlet should be ignored for dynamic purposes given it only throws if metadata itself was dynamic.

This change fixes the order, prioritizing Suspense behind the special case boundaries.