fix(after): broken error handling for unstable_after(promise) (#71373)
In #71231 we've added `onAfterTaskError`, which is intended for handling
errors that happened within `unstable_after` (currently only used to
fail the build if anything is thrown during prerendering). It was being
called correctly for the callback form, `unstable_after(() => { ... })`,
but we missed the fact that `unstable_after(promise)` is also valid, and
weren't calling it for rejected promises.

(Apparently we also weren't printing any kind of error message for
those, and it's been that way since the start. my bad.)

This PR unifies the error handling for callbacks and promises.