[dynamicIO] track environment on requestStore  (#71345)
Stacked on #71344 

environment is used to contextualize logs. We use the Prerender
environment in the first task of a render and Server in the remaining
tasks. This was previously implemented using a closed over variable but
we actually want to be able to make error determinations in dev mode
using this information so in this change we move this to the
requeststore