Increase timeout for deploy tests from 120 (default) to 240 seconds (#71428)
When running deploy tests, the deployments are done in the `beforeAll`
hook of a test. For some more complex test apps the deployment may take
a while to complete which can result in the test timing out after 120
seconds. To mitigate this, we increase the timeout to 240 seconds.

x-ref: [example run with a
timeout](https://github.com/vercel/next.js/actions/runs/11385828170/job/31678217067)