Let cache body execute with hanging inputs (#71452)
This lets us render deeper into the "use cache" entry as if it wasn't
there for the purposes of prerendering.

One such hanging input is the params/searchParams passed to pages.

For now we error if you end up using the dynamic input after a timeout.
We could generate a useful partially dynamic output here if we generated
a halted RSC payload but for now we just error. Doing this without a
timeout is tricky.