[Turbopack] add a startup cache which caches the first 100MB of accessed data (#71431)
### What?

When the OS filesystem cache is cold, accessing the database can be
quite slow and add a few seconds of build time since database is
accessed kind of randomly.

This adds a cache file on top of the key value database which caches the
first 100MB of data accessed from the database and puts it into a
separate file. That file is sequentially read and pulled into memory on
startup.