[dynamicIO] Move disallowed dynamic tracking to it's own object (#71499)
dynamicTracking is currently implemented as a unififed object that is
shared across the RSC/SSR renders. Except for the newly added disallowed
dynamic tracking it is essentially unused for dynamicIO (though we
technically haven't removed some lesser used debugging tools originally
added for PPR when that is enabled). This change extracts the disallowed
dynamic state into it's own object on the prerender store. We do this so
we can model it as single-shot. i.e you don't have one for RSC and SSR
that is shared you just have one for SSR (that's the only place we can
track disallowed dynamic anyway).

This sets the state for prospective SSR renders