[dynamicIO] show dev & build warnings when using incompatible segment configs (#71500)
This logs an error in dev and fails a build when using incompatible
segment configs (`dynamic`, `fetchCache`, `runtime`, `dynamicParams`)
with dynamicIO. It will point to all discovered pages that contained the
invalid segment config at the end of the build.

![CleanShot 2024-10-18 at 15 55
17@2x](https://github.com/user-attachments/assets/d8c21ede-a71d-4b26-a76d-26e330aa8de2)

In dev, it will show the error in an overlay. The invalid segment
configs must be removed to build.

![CleanShot 2024-10-19 at 11 47
59@2x](https://github.com/user-attachments/assets/c53797d8-5729-4567-9fb0-5d12c2860382)

Note: This removes all incompatible test cases that were previously part
of the `dynamic-io` test suites, and disables one that was using a
fetchCache workaround until it's properly handled.