Avoid server action function indirection in Turbopack
As a follow-up to #71572, this PR replaces the server action function
indirection in Turbopack's generated action loader module with a direct
re-export of the actions, using the action ID as export name.

The server action identity test that was added in #71572 was not
sufficient to uncover that Turbopack was still using the wrapper
functions. To fix that, we've added another test here, which uses a
combination of `"use server"` and `"use cache"` functions in a page.
This test does fail in Turbopack without the loader changes.

In addition, the building of the server reference manifest is adjusted
to account for the new structure that expects `{moduleId: string, async:
boolean}` instead of `string`. The `async` flag is set based on the
loader's chunk item. I believe this is currently always `false` though.
However, Turbopack can still resolve the modules just fine, even when
there are outgoing async connections.