Implement information byte in Server Reference ID and other optimizations (#71463)
With this PR, we're adding one extra leading byte to Server Reference
IDs (both Server Actions and `"use cache"` functions), to include some
static information about the function itself.

The information byte has the following format:

```
0     000000    0
^type ^arg mask ^rest args
```

The type bit represents if the action is a cache function or not. For
cache functions, the type bit is set to `1`. Otherwise, it's `0`.

The arg mask bit is used to determine which arguments are used by the
function itself, up to 6 arguments. The bit is set to `1` if the
argument is used, or being spread or destructured (so it can be
indirectly or partially used). The bit is set to `0` otherwise.

The rest args bit is used to determine if there's a `...` rest argument
in the function signature. If there is, the bit is set to `1`.

For example:

```tsx
async function foo(a, foo, b, bar, ...baz) {
  'use cache';
  return a + b;
}
```

will have it encoded as `[1][101011][1]`. The first bit is set to `1`
because it's a cache function. The second part has `1010` because the
only arguments used are `a` and `b`. The subsequent `11` bits are set to
`1` because there's a `...baz` argument starting from the 5th. The last
bit is set to `1` as well for the same reason.

Note: Currently in this PR we don't track if an argument is actually
referenced in the function body or not. That will be implemented as a
follow-up optimization.

Also, the reference ID is currently hex-encoded so there will be exact 2
characters for easier decoding. This encoding might change though.

With this extra byte, the client can do some further optimizations. More
details can be found in the code comment.