Use `registerClientReference` for ESM client component modules
When registering a client reference for a client component module, React
offers two distinct functions:

- `registerClientReference`: Register a named or default export of an
  ES module as a client reference.
- `createClientModuleProxy`: Create a client module proxy for a CJS
  module that registers every module export as a client reference.

Until this PR, we were using the client module proxy for both kinds of
modules, which is incorrect. It was especially problematic because in
the Node.js runtime we were always awaiting the client module proxy.
This was a workaround for handling async modules (see #66990), but led
to issues like #71840.

With this PR, we are now using `registerClientReference` if we detect
that a module is an ES module. This aligns the Webpack implementation
with the one in Turbopack.

fixes #71840