Prevent prerender errors due to `Math.random()` usage by `@vercel/otel`
When [`@vercel/otel` is configured as OpenTelemetry
implementation](https://nextjs.org/docs/app/building-your-application/optimizing/open-telemetry#using-vercelotel),
and `experimental.dynamicIO` is enabled, an error is logged during
prerendering with `next dev`, due to the usage of `Math.random()` for
generating span IDs. This is triggered by the internal tracing of
Next.js.

To fix this, we're exiting the work unit storage before creating a span.
This is a solution we're already using in `patchErrorInspect` for the
same reason (in this case triggered by the `source-map` library).

Notably, this does not prevent the errors from being logged when app
code contains trace calls. For now, users need to use either `await
connection()` or `"use cache"` when tracing with `@vercel/otel`.

fixes #72024